% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_methods.R
\name{imp_methods}
\alias{imp_methods}
\title{Get available imputation methods based on data type}
\usage{
imp_methods(data_type = "binary")
}
\arguments{
\item{data_type}{The type of data in the outcome column (supported data types: "binary", "numeric", "ordinal", "nominal")}

\item{level}{The level of imputation method ("1" or "2")}
}
\value{
A character vector of available imputation methods from the mice package, including their full name and mice code
}
\description{
Get available imputation methods based on data type
}
\examples{
methods <- imp_methods("binary")
}
