% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_gone.R
\name{missing_gone}
\alias{missing_gone}
\title{remove missing data based on thresholds created by user}
\usage{
missing_gone(data, threshold = 0.1, exclude = NULL)
}
\arguments{
\item{data}{A data frame that has the missing data}

\item{threshold}{The threshold that represents the percentage of missing data in the variable.
Variables with missing data percent frequency above this threshold will be removed.}
}
\value{
A data frame with variables removed based on the threshold
}
\description{
remove missing data based on thresholds created by user
}
\examples{
Using the "here" package to locate the example file:
library(here)
data <- here("data-raw", "conflict_data.csv")
Write in the function:
data <- missing_gone(data, threshold = 0.1, exclude = "Maternal Mortality Rate")
}
