% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNAR.R
\name{test_missing_at_random}
\alias{test_missing_at_random}
\title{Test for Missing at Random}
\usage{
test_missing_at_random(data, p_value = 0.05)
}
\arguments{
\item{data}{A data frame with missing data.}

\item{p_value}{The significance level for the logistic regression tests.}
}
\value{
A data frame summarizing the results of logistic regression tests.
}
\description{
This function tests whether missing data is at random using logistic regression.
}
