% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_methods.R
\name{imp_methods}
\alias{imp_methods}
\title{Get available imputation methods based on data type}
\usage{
imp_methods(data_type, level)
}
\arguments{
\item{data_type}{The type of data in the outcome column (eg. "binary", "numeric", "ordinal", "nominal")}

\item{methods}{Built-in uni-variate imputation methods from the mice package}
}
\value{
A character vector of available imputation methods
}
\description{
Get available imputation methods based on data type
}
\examples{
methods <- imp_methods("binary")
}
